031020-13:10
Replaced onAction() with onActionBegin() to prevent a ball from being spawned very frame.

Added a function in BallComponent called initVelocity(), which takes a double and uses it to initialize each instance's vertical velocity via their PhysicsComponents.

Added a removeFromWorld() call in PongApp which is called on each ball instance when it strikes a player.